+++ packages/botonic-plugin-flow-builder/src/types.ts
@@ -2,7 +2,6 @@ import { PluginPreRequest, Session } from '@botonic/core'
 import { ActionRequest } from '@botonic/react'
 
 import { HtFlowBuilderData } from './content-fields/hubtype-fields'
-import { InferenceParams } from './user-input/smart-intent'
 
 export interface BotonicPluginFlowBuilderOptions {
   apiUrl?: string